Function random::init_rng

pub fn init_rng<T>() -> Result<T, ()>where
    T: SeedableRng,

Initialises a T RNG.

Directly accessing the global CSPRNG can be expensive and so it is often better to seed a local PRNG from the global CSPRNG. Using a local PRNG also allows for much faster cryptographically insecure PRNGs to be used.