Struct random::Default

pub struct Default(_);

An instance of the default source.

The default source is the Xorshift128+ algorithm.

Methods

impl Default

fn new() -> Default

Create an instance of the default source.

Each thread has its own copy of the default source, and each copy is initialized with the same default seed. Therefore, the usage is thread safe; however, each thread is responsible for reseeding its source.

fn seed(self, seed: [u64; 2]) -> Default

Seed the source.

At least one bit of the seed should be nonzero.