Crate random

Re-exports

pub use weighted_bag::WeightedBag;bag

Modules

weighted_bagbag

Functions

conflip

Returns true 50% of the time

get

Samples a number from the range x..y

get_inc

Samples a number from the range x..=y

pick

Select a random element from the given slice

seed

Retrieves the seed

set_seed

Sets the seed for the future queries This is mostly usefull to make deterministic tests for games, or even bug hunts

str

Samples a random String with a given lengh